move device_manager fixture to conftest
